backport some fixes find in main repo

+++ MQTTPacket/src/MQTTPacket.h
@@ -35,7 +35,7 @@ DLLExport int32_t MQTTSerialize_ack(unsigned char* buf, size_t buflen, unsigned 
 DLLExport int32_t MQTTDeserialize_ack(unsigned char* packettype, unsigned char* dup, unsigned short* packetid, unsigned char* buf, size_t buflen);
 
 DLLExport int MQTTPacket_equals(const MQTTString* a, char* b);
-DLLExport int32_t MQTTPacket_encode(unsigned char* buf, int32_t length);
+DLLExport int32_t MQTTPacket_encode(unsigned char* buf, size_t length);
 DLLExport int MQTTPacket_read(unsigned char* buf, size_t buflen, int (*getfn)(unsigned char*, int));
 typedef struct {
 	int (*getfn)(void *, unsigned char*, int); /* must return -1 for error, 0 for call again, or the number of bytes read */
